feat: 채널 아이디 중복 체크 / 채널 생성
