+++ src/reducers/index.js
@@ -2,13 +2,10 @@ import { combineReducers } from 'redux';
 import user from './user';
 import post from './post';
 import channel from './channel';
+import hostcenter from './hostcenter';
 
 const initialState = {
   status: 'test',
-  hostcenter: {
-    isDuplicated: null,
-    currentChannel: null,
-  },
 };
 
 const rootReducer = combineReducers({
@@ -17,22 +14,6 @@ const rootReducer = combineReducers({
       case 'TEST':
         console.log('test');
         return state;
-      case 'URL_DUPLICATE_CHECK_REQUEST':
-        return {
-          ...state,
-          hostcenter: {
-            ...state.hostcenter,
-            isDuplicated: true,
-          },
-        };
-      case 'CREATE_CHANNEL_REQUEST':
-        return {
-          ...state,
-          hostcenter: {
-            ...state.hostcenter,
-            currentChannel: action.data.channelID,
-          },
-        };
       default:
         return state;
     }
@@ -40,6 +21,7 @@ const rootReducer = combineReducers({
   user,
   post,
   channel,
+  hostcenter,
 });
 
 export default rootReducer;
