+++ src/sagas/index.js
@@ -3,6 +3,7 @@ import axios from 'axios';
 import channelSaga from './channel';
 import userSaga from './user';
 import postSaga from './post';
+import hostcenterSaga from './hostcenter';
 
 axios.defaults.withCredentials = true;
 
@@ -11,5 +12,6 @@ export default function* rootSaga() {
     fork(channelSaga),
     fork(userSaga),
     fork(postSaga),
+    fork(hostcenterSaga),
   ]);
 }
